+++ plugins/processing/signal-processing/src/box-algorithms/basic/ovpCBoxAlgorithmChannelRename.cpp
@@ -60,7 +60,7 @@ bool CBoxAlgorithmChannelRename::uninitialize()
 	return true;
 }
 
-bool CBoxAlgorithmChannelRename::processInput(const uint32_t /*index*/)
+bool CBoxAlgorithmChannelRename::processInput(const size_t /*index*/)
 {
 	this->getBoxAlgorithmContext()->markAlgorithmAsReadyToProcess();
 	return true;
