+++ plugins/processing/signal-processing/src/box-algorithms/basic/ovpCBoxAlgorithmCrop.cpp
@@ -82,7 +82,7 @@ bool CBoxAlgorithmCrop::uninitialize()
 	return true;
 }
 
-bool CBoxAlgorithmCrop::processInput(const uint32_t /*index*/)
+bool CBoxAlgorithmCrop::processInput(const size_t /*index*/)
 {
 	getBoxAlgorithmContext()->markAlgorithmAsReadyToProcess();
 	return true;
