+++ plugins/processing/signal-processing/src/box-algorithms/basic/ovpCBoxAlgorithmEpochAverage.cpp
@@ -93,7 +93,7 @@ bool CBoxAlgorithmEpochAverage::uninitialize()
 	return true;
 }
 
-bool CBoxAlgorithmEpochAverage::processInput(const uint32_t /*index*/)
+bool CBoxAlgorithmEpochAverage::processInput(const size_t /*index*/)
 {
 	getBoxAlgorithmContext()->markAlgorithmAsReadyToProcess();
 	return true;
