+++ plugins/processing/signal-processing/src/box-algorithms/basic/ovpCBoxAlgorithmIdentity.h
@@ -11,7 +11,7 @@ namespace OpenViBEPlugins
 		{
 		public:
 			void release() override;
-			bool processInput(const uint32_t index) override;
+			bool processInput(const size_t index) override;
 			bool process() override;
 
 			_IsDerivedFromClass_Final_(OpenViBEToolkit::TBoxAlgorithm<OpenViBE::Plugins::IBoxAlgorithm>, OVP_ClassId_BoxAlgorithm_Identity)
