+++ plugins/processing/signal-processing/src/box-algorithms/epoching/ovpCBoxAlgorithmStimulationBasedEpoching.cpp
@@ -60,7 +60,7 @@ bool CBoxAlgorithmStimulationBasedEpoching::uninitialize()
 	return true;
 }
 
-bool CBoxAlgorithmStimulationBasedEpoching::processInput(const uint32_t /*index*/)
+bool CBoxAlgorithmStimulationBasedEpoching::processInput(const size_t /*index*/)
 {
 	this->getBoxAlgorithmContext()->markAlgorithmAsReadyToProcess();
