+++ plugins/processing/signal-processing/src/box-algorithms/resampling/ovpCBoxAlgorithmSignalResampling.cpp
@@ -87,7 +87,7 @@ bool CBoxAlgorithmSignalResampling::uninitialize()
 	return true;
 }
 
-bool CBoxAlgorithmSignalResampling::processInput(const uint32_t /*index*/)
+bool CBoxAlgorithmSignalResampling::processInput(const size_t /*index*/)
 {
 	this->getBoxAlgorithmContext()->markAlgorithmAsReadyToProcess();
 	return true;
