+++ plugins/processing/signal-processing/src/box-algorithms/spectral-analysis/ovpCBoxAlgorithmFrequencyBandSelector.cpp
@@ -115,7 +115,7 @@ bool CBoxAlgorithmFrequencyBandSelector::uninitialize()
 	return true;
 }
 
-bool CBoxAlgorithmFrequencyBandSelector::processInput(const uint32_t /*index*/)
+bool CBoxAlgorithmFrequencyBandSelector::processInput(const size_t /*index*/)
 {
 	getBoxAlgorithmContext()->markAlgorithmAsReadyToProcess();
 	return true;
