+++ plugins/processing/tools/src/box-algorithms/ovpCBoxAlgorithmMatrixValidityChecker.cpp
@@ -60,7 +60,7 @@ bool CBoxAlgorithmMatrixValidityChecker::uninitialize()
 	return true;
 }
 
-bool CBoxAlgorithmMatrixValidityChecker::processInput(const uint32_t /*index*/)
+bool CBoxAlgorithmMatrixValidityChecker::processInput(const size_t /*index*/)
 {
 	getBoxAlgorithmContext()->markAlgorithmAsReadyToProcess();
 	return true;
