+++ plugins/processing/tools/src/box-algorithms/ovpCBoxAlgorithmStimulationListener.cpp
@@ -33,7 +33,7 @@ bool CBoxAlgorithmStimulationListener::uninitialize()
 	return true;
 }
 
-bool CBoxAlgorithmStimulationListener::processInput(const uint32_t /*index*/)
+bool CBoxAlgorithmStimulationListener::processInput(const size_t /*index*/)
 {
 	getBoxAlgorithmContext()->markAlgorithmAsReadyToProcess();
 	return true;
